finished getting started polishes
