+++ Labs/Introduction/GettingStarted/testDriver.py
@@ -1,0 +1,216 @@
+# solutions.py
+"""Introductory Labs: Getting Started. Test Driver."""
+
+from numpy.random import randint
+from solutions import *
+
+# Test script        
+def test(student_module):
+    """Test script. Import the student's solutions file as a module.
+    
+     5 points for problem 2
+    10 points for problem 3
+     5 points for problem 4
+     5 points for problem 5
+     5 points for problem 6
+    10 points for problem 8
+    
+    Parameters:
+        student_module: the imported module for the student's file.
+    
+    Returns:
+        score (int): the student's score, out of 40.
+        feedback (str): a printout of results for the student.
+    """
+    tester = _testDriver()
+    tester.test_all(student_module)
+    return tester.score, tester.feedback
+
+class _testDriver(object):
+    """Class for testing a student's work. See test.__doc__ for more info.
+
+    This and all other test drivers can be used to grade the entire lab
+    assignment at once via test_all(), or to grade one problem at a time
+    via the different problemX() functions.
+
+    The point distribution is only a suggestion; the instructor may alter
+    the weight of each problem as they see fit.
+    """
+
+    # Constructor -------------------------------------------------------------
+    def __init__(self):
+        """Initialize the feedback attribute."""
+        self.feedback = ""
+
+    # Main routine -----------------------------------------------------------
+    def test_all(self, student_module, total=40):
+        """Grade the provided module on each problem and compile feedback."""
+        # Reset feedback and score.
+        self.feedback = ""
+        self.score = 0
+
+        def test_one(problem, label, value):
+            """Test a single problem, checking for errors."""
+            try:
+                self.feedback += "\n\n{} ({} points):".format(label, value)
+                points = problem(student_module)
+                self.score += points
+                self.feedback += "\nScore += {}".format(points)
+            except BaseException as e:
+                self.feedback += "\n{}: {}".format(self._errType(e),e)
+
+        # Grade each problem.
+        test_one(self.problem2, "Problem 2",  5)    # Problem 2:  5 points.
+        test_one(self.problem3, "Problem 3", 10)    # Problem 3: 10 points.
+        test_one(self.problem4, "Problem 4",  5)    # Problem 4:  5 points.
+        test_one(self.problem5, "Problem 5",  5)    # Problem 5:  5 points.
+        test_one(self.problem6, "Problem 6",  5)    # Problem 6:  5 points.
+        test_one(self.problem7, "Problem 7", 10)    # Problem 7: 10 points.
+
+        # Report final score.
+        percentage = (100. * self.score) / total
+        self.feedback += "\n\nTotal score: {}/{} = {}%".format(
+                                    self.score, total, round(percentage, 2))
+        if   percentage >=  98: self.feedback += "\n\nExcellent!"
+        elif percentage >=  90: self.feedback += "\n\nGreat job!"
+
+        # Add comments (optionally).
+        print(self.feedback)
+        comments = str(raw_input("Comments: "))
+        if len(comments) > 0:
+            self.feedback += '\n\n\nComments:\n\t{}'.format(comments)
+
+    # Helper Functions --------------------------------------------------------
+    @staticmethod
+    def _errType(error):
+        """Get just the name of the exception 'error' in string format."""
+        return str(type(error).__name__)
+
+    def _eqTest(self, correct, student, message):
+        """Test to see if 'correct' and 'student' are equal.
+        Report the given 'message' if they are not.
+        """
+        if correct == student:
+            return 1
+        else:
+            self.feedback += "\n{}".format(message)
+            self.feedback += "\n\tCorrect response: {}".format(correct)
+            self.feedback += "\n\tStudent response: {}".format(student)
+            return 0
+
+    def _grade(self, points, message=None):
+        """Manually grade a problem worth 'points'. Return the score.
+        If full points are not earned, get feedback on the problem.
+        """
+        credit = -1
+        while credit > points or credit < 0:
+            try:
+                credit = int(input("\nScore out of {}: ".format(points)))
+            except:
+                credit = -1
+        if credit != points:
+            # Add comments (optionally),
+            comments = raw_input("Comments: ")
+            if len(comments) > 0:
+                self.feedback += "\n{}".format(comments)
+            # Or add a predetermined error message.
+            elif message is not None:
+                self.feedback += "\n{}".format(message)
+        return credit
+
+
+    # Problems ----------------------------------------------------------------
+    def problem2(self, s):
+        """Test sphere_volume(). 5 Points."""
+        if not hasattr(s, "sphere_volume"):
+            raise NotImplementedError("Problem 2 Incomplete")
+        if s.sphere_volume(1) is None:
+            raise NotImplementedError("sphere_volume() returned nothing")
+        
+        points  = 2*self._eqTest(sphere_volume(5), s.sphere_volume(5),
+                                            "sphere_volume(5) failed")
+        points += 3*self._eqTest(sphere_volume(3.14), s.sphere_volume(3.14),
+                                            "sphere_volume(3.14) failed")
+        return points
+
+    def problem3(self, s):
+        """Test first_half() and backward(). 10 points."""
+        if not hasattr(s, "first_half") or not hasattr(s, "backward"):
+            raise NotImplementedError("Problem 3 Incomplete")
+        if s.first_half("abcde") is None:
+            raise NotImplementedError("first_half() returned nothing")
+        if s.backward("abcde") is None:
+            raise NotImplementedError("backward() returned nothing")
+
+        points  = 2*self._eqTest(first_half("abcde"), s.first_half("abcde"),
+                                            "first_half('abcde') failed")
+        points += 3*self._eqTest(first_half("TK421"), s.first_half("TK421"),
+                                            "first_half('TK421') failed")
+        points += 2*self._eqTest(backward("abcde"), s.backward("abcde"),
+                                            "first_half('abcde') failed")
+        points += 3*self._eqTest(backward("TK421"), s.backward("TK421"),
+                                            "first_half('TK421') failed")
+        return points
+
+    def problem4(self, s):
+        """Test list_ops(). 5 points."""
+        if not hasattr(s, "list_ops"):
+            raise NotImplementedError("Problem 4 Incomplete")
+        if s.list_ops(["ant", "bear", "cat", "dog"]) is None:
+            raise NotImplementedError("list_ops() returned nothing")
+        
+        points = 2*self._eqTest(  list_ops(["ant", "bear", "cat", "dog"]),
+                                s.list_ops(["ant", "bear", "cat", "dog"]),
+                        'list_ops(["ant", "bear", "cat", "dog"]) failed')
+        points +=3*self._eqTest(list_ops(["ant", "beaver", "cobra", "dragon"]),
+                            s.list_ops(["ant", "beaver", "cobra", "dragon"]),
+                'list_ops(["ant", "beaver", "cobra", "dragon"]) failed')
+        return points
+
+    def problem5(self, s):
+        """Test pig_latin(). 5 points."""
+        if not hasattr(s, "pig_latin"):
+            raise NotImplementedError("Problem 5 Incomplete")
+        if s.pig_latin("piglatin") is None:
+            raise NotImplementedError("pig_latin() returned nothing")
+
+        points = 2*self._eqTest(  pig_latin("college"),
+                                s.pig_latin("college"),
+                                            "pig_latin('college') failed")
+        points += 3*self._eqTest(  pig_latin("university"),
+                                 s.pig_latin("university"),
+                                        "pig_latin('university') failed")
+        return points
+
+    def problem6(self, s):
+        """Test int_to_string(). 5 points.""" 
+        if not hasattr(s, "int_to_string"):
+            raise NotImplementedError("Problem 6 Incomplete")
+        if s.int_to_string([1]) is None:
+            raise NotImplementedError("int_to_string() returned nothing")
+        
+        points = 2*self._eqTest(int_to_string([1, 2, 3]),
+                                s.int_to_string([1, 2, 3]),
+                                    "int_to_string([1, 2, 3]) failed")
+        points += 3*self._eqTest(
+                        int_to_string([24, 25, 12, 15, 16, 8, 15, 14, 5]),
+                        s.int_to_string([24, 25, 12, 15, 16, 8, 15, 14, 5]),
+            "int_to_string([24, 25, 12, 15, 16, 8, 15, 14, 5]) failed")
+        return points
+
+    def problem7(self, s):
+        """Test alt_harmonic(). 10 points."""
+        if not hasattr(s, "alt_harmonic"):
+            raise NotImplementedError("Problem 8 Incomplete")
+        if s.alt_harmonic(10) is None:
+            raise NotImplementedError("alt_harmonic() returned nothing")
+
+        points = 4*self._eqTest(alt_harmonic(100), s.alt_harmonic(100),
+                                            "alt_harmonic(100) failed")
+        points += 6*self._eqTest(alt_harmonic(5000), s.alt_harmonic(5000),
+                                            "alt_harmonic(5000) failed")
+        return points
+
+if __name__ == '__main__':
+    import solutions as sol
+    score, feedback = test(sol)
